+++ monarch/lib/src/builders/meta_locales_generator.dart
@@ -4,13 +4,13 @@ import 'package:analyzer/dart/element/element.dart';
 
 import 'package:monarch_annotations/monarch_annotations.dart';
 
-const TypeChecker monarchLocalizationTypeChecker =
-    TypeChecker.fromRuntime(MonarchLocalization);
+const TypeChecker monarchLocalizationsTypeChecker =
+    TypeChecker.fromRuntime(MonarchLocalizations);
 
 class MetaLocalizationsGenerator extends Generator {
   @override
   String generate(LibraryReader library, BuildStep buildStep) {
-    final annotations = library.annotatedWith(monarchLocalizationTypeChecker);
+    final annotations = library.annotatedWith(monarchLocalizationsTypeChecker);
     if (annotations.isEmpty) {
       return null;
     }
@@ -20,30 +20,29 @@ class MetaLocalizationsGenerator extends Generator {
     for (var annotatedElement in annotations) {
       final element = annotatedElement.element;
 
-      if (element is ClassElement) {
+      if (element is TopLevelVariableElement) {
         log.fine(
-            'Found MonarchLocalization annotation on class element: ${element.name}');
+            'Found MonarchLocalizations annotation on top-level element: ${element.name}');
         final annotation = annotatedElement.annotation;
 
         final localesDartObjects = annotation.read('locales').listValue;
-        final localeCodes = localesDartObjects.map((x) => x.toStringValue());
-        final locales = localeCodes.map((x) => "Locale('$x')");
+        final locales = localesDartObjects.map((x) {
+          final languageCode = x.getField('languageCode').toStringValue();
+          final countryCode = x.getField('countryCode').toStringValue();
+          if (countryCode == null) {
+            return "Locale('$languageCode')";
+          } else {
+            return "Locale('$languageCode', '$countryCode')";
+          }
+        });
         final localesExpression = "[${locales.join(', ')}]";
 
-        if (element.supertype.typeArguments.isEmpty) {
-          log.warning(
-              'Superclass of ${element.name} does not have parameter type T. Skipping.');
-        } else {
-          final typeParameterName =
-              element.supertype.typeArguments[0].element.name;
-
-          expressions.add(
-              "MetaLocalization<$typeParameterName>.user($localesExpression, ${element.name}(), '${element.name}')");
-        }
+        expressions.add(
+            "MetaLocalization.user($localesExpression, ${element.name}, '${element.name}')");
       } else {
         final msg = '''
-Found MonarchLocalization annotation on an element that is not a class declaration.
-The MonarchLocalization annotation must be placed on a class declaration.
+Found MonarchLocalizations annotation on an element that is not a top-level variable.
+The MonarchLocalizations annotation must be placed on a top-level variable.
 Element name: ${element.name}
 ''';
         log.warning(msg);
@@ -54,18 +53,18 @@ Element name: ${element.name}
       return _outputContents(buildStep.inputId.uri.toString(), expressions);
     } else {
       log.warning(
-          'Could not compute import URI to file with MonarchLocalization annotation. '
+          'Could not compute import URI to file with MonarchLocalizations annotation. '
           'Please make sure the file is inside the lib directory.');
       return null;
     }
   }
 
-  String _outputContents(
-      String pathToLocalizationFile, List<String> metaLocalizationExpressions) {
+  String _outputContents(String pathToLocalizationsFile,
+      List<String> metaLocalizationExpressions) {
     return '''
 import 'dart:ui';
 import 'package:monarch/monarch.dart';
-import '$pathToLocalizationFile';
+import '$pathToLocalizationsFile';
 
 final metaLocalizationItems = [
   ${metaLocalizationExpressions.join(', ')}